Minor tweak
