+++ libs/DocumentScanner/src/main/java/com/zynksoftware/documentscanner/ui/base/BaseFragment.kt
@@ -19,13 +19,33 @@
 
 package com.zynksoftware.documentscanner.ui.base
 
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
 import android.widget.RelativeLayout
 import androidx.fragment.app.Fragment
+import androidx.viewbinding.ViewBinding
 import com.zynksoftware.documentscanner.R
 import com.zynksoftware.documentscanner.common.extensions.hide
 import com.zynksoftware.documentscanner.common.extensions.show
 
-internal abstract class BaseFragment : Fragment() {
+internal abstract class BaseFragment<BINDING : ViewBinding>(private val bindingInflater: (layoutInflater: LayoutInflater) -> BINDING) : Fragment() {
+
+    private var _binding: BINDING? = null
+
+    // This property is only valid between onCreateView and onDestroyView.
+    val binding get() = _binding!!
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        _binding = bindingInflater(inflater)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
 
     fun showProgressBar() {
         view?.findViewById<RelativeLayout>(R.id.progressLayout)?.show()
